requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-JWT-Extended>=4.6
Flask-Login>=0.6
Flask-Limiter>=3.5
Flask-WTF>=1.2
scikit-learn>=1.4
xgboost>=2.0
spacy>=3.7
nltk>=3.8
vaderSentiment>=3.3
textstat>=0.7
pandas>=2.1
numpy>=1.26
joblib>=1.3
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
Werkzeug>=3.0
reportlab>=4.0
email-validator>=2.1
# PostgreSQL driver for Supabase
psycopg2-binary>=2.9
# Supabase Python SDK
supabase>=2.3
